separate report_data for the render_report test
